\name{choixvarfunc}
\alias{choixvarfunc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Selection of a variable before plotting bubbles}

\description{
The function \code{choixvarfunc()} is used in the function \code{choix.bubble()}.
}

\usage{
choixvarfunc(title, question, liste)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
   \item{title}{a character as "Choice of variables"}
   \item{question}{a character as "Choose a variable"}
   \item{liste}{a list of character with names of variables}
}
\details{This function is not an interactive function.}

\value{Name of variable chosen.}
\references{Aragon Yves, Perrin Olivier, Ruiz-Gazen Anne, Thomas-Agnan Christine (2010), \emph{Statistique et Econométrie pour données géoréférencées : modèles et études de cas}}
\author{Laurent T.}

\keyword{spatial}
\keyword{utilities} 
