\name{meuse.riv}
\alias{meuse.riv}
\docType{data}
\title{River Meuse outline}
\description{
The `meuse.riv' data consists of an outline of the Meuse river in the area a few kilometers around the meuse data set.
}
\usage{data(meuse.riv)}
\format{
This data frame contains a $176 \times 2$ matrix with coordinates.
}
\details{
This data set provides from spdep package. 
}
\source{
The actual field data were collected by Ruud van Rijn and Mathieu Rikken; data compiled for R by Edzer J. Pebesma

}
\references{
P.A. Burrough, R.A. McDonnell, 1998. Principles of Geographical Information Systems. Oxford University Press, \url{http:/www.gstat.org/}
}
\examples{
data(meuse.riv)
}
\keyword{datasets}
