\name{selectgraph}
\alias{selectgraph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Selection of an additionnal grah}

\description{
The function \code{selectgraph()} is used in most of the GeoXp functions to choose an additional graph and the variables associated to this graph.
}

\usage{
selectgraph(listnomvar, listgraph)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
   \item{listnomvar}{list of names of variables given}
   \item{listgraph}{kind of graphics }
}
\details{This function is not an interactive function.}

\value{Names of variables and graph chosen.}
\references{Aragon Yves, Perrin Olivier, Ruiz-Gazen Anne, Thomas-Agnan Christine (2009), \emph{Statistique et Econométrie pour données géoréférencées : modèles et études de cas}}
\author{Thomas-Agnan C., Aragon Y., Ruiz-Gazen A., Laurent T., Robidou L.}

\keyword{spatial}
\keyword{utilities} 
