\name{choix.bubble}
\alias{choix.bubble}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Selection of a variable before plotting bubbles}

\description{
The function \code{choix.bubble()} is used in most of the GeoXp functions to initialize some parameters before plotting bubbles.
}

\usage{
choix.bubble(buble,listvar,listnomvar,legends)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
   \item{buble}{boolean}
   \item{listvar}{list of variables}
   \item{listnomvar}{names of variables}
   \item{legends}{parameters of plot}
}

\details{This function is not an interactive function.}

\value{A list of parameters which permit to use the function carte.}
\references{Aragon Yves, Perrin Olivier, Ruiz-Gazen Anne, Thomas-Agnan Christine (2009), \emph{Statistique et Econométrie pour données géoréférencées : modèles et études de cas}}
\author{Laurent T.}

\keyword{spatial}
\keyword{utilities} 
