\name{choixvarfunc}
\alias{choixvarfunc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Selection of a variable before plotting bubbles}

\description{
The function \code{choixvarfunc()} is used in the function \code{choix.bubble()}.
}

\usage{
choixvarfunc(title, question, liste)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
   \item{title}{a character as "Choice of variables"}
   \item{question}{a character as "Choose a variable"}
   \item{liste}{a list of character with names of variables}
}
\details{This function is not an interactive function.}

\value{Name of variable chosen.}
\references{Thibault Laurent, Anne Ruiz-Gazen, Christine Thomas-Agnan (2012), GeoXp: An R Package for Exploratory Spatial Data Analysis. \emph{Journal of Statistical Software}, 47(2), 1-23.}
\author{Laurent T.}

\keyword{spatial}
\keyword{utilities} 
