\name{selectgraph}
\alias{selectgraph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Selection of an additionnal grah}

\description{
The function \code{selectgraph()} is used in most of the GeoXp functions to choose
an additional graph and the variables associated to this graph.
}

\usage{
selectgraph(listnomvar, listgraph)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
   \item{listnomvar}{list of names of variables given}
   \item{listgraph}{kind of graphics }
}
\details{This function is not an interactive function.}

\value{Names of variables and graph chosen.}
\references{Thibault Laurent, Anne Ruiz-Gazen, Christine Thomas-Agnan (2012), GeoXp: An R Package for Exploratory Spatial Data Analysis. \emph{Journal of Statistical Software}, 47(2), 1-23}
\author{Thomas-Agnan C., Aragon Y., Ruiz-Gazen A., Laurent T., Robidou L.}

\keyword{spatial}
\keyword{utilities} 
